Refactoring - Create new unit test modules for each new module, rearrange tests accordingly.
