+++ tests/unit/test_message_utilities.py
@@ -1,7 +1,10 @@
 """Unit tests for message_utilities.py."""
 
+import pytest
 from harmony_service_lib.message import Message as HarmonyMessage
 
+from harmony_regridding_service.exceptions import InvalidTargetCRS
+from harmony_regridding_service.message_utilities import has_valid_crs
 from harmony_regridding_service.regridding_cli import (
     get_harmony_message_from_params,
 )
@@ -56,3 +59,56 @@ def test_get_harmony_message_no_params():
     assert message.format.scaleSize is None
     assert message.format.height is None
     assert message.format.width is None
+
+
+@pytest.mark.parametrize(
+    'message, expected, description',
+    [
+        (HarmonyMessage({}), True, 'CRS = None is valid'),
+        (HarmonyMessage({'format': {}}), True, 'format.crs = None is valid'),
+        (
+            HarmonyMessage({'format': {'crs': 'EPSG:4326'}}),
+            True,
+            'format.crs = "EPSG:4326" is valid',
+        ),
+        (
+            HarmonyMessage({'format': {'crs': '+proj=longlat'}}),
+            True,
+            'format.crs = "+proj=longlat" is valid',
+        ),
+        (
+            HarmonyMessage({'format': {'crs': '4326'}}),
+            True,
+            'format.crs = "4326" is valid',
+        ),
+        (
+            HarmonyMessage({'format': {'crs': 'EPSG:6933'}}),
+            False,
+            'Non-geographic EPSG code is invalid',
+        ),
+        (
+            HarmonyMessage({'format': {'crs': '+proj=cea'}}),
+            False,
+            'Non-geographic proj4 string is invalid',
+        ),
+    ],
+)
+def test_has_valid_crs(message, expected, description):
+    """Test has_valid_crs.
+
+    Ensure the function correctly determines if the input Harmony
+    message has a target Coordinate Reference System (CRS) that is
+    compatible with the service. Currently this is either to not
+    define the target CRS (assuming it to be geographic), or explicitly
+    requesting geographic CRS via EPSG code or proj4 string.
+
+    """
+    assert has_valid_crs(message) == expected, f'Failed for {description}'
+
+
+def test_has_valid_crs_raises_exception():
+    """Test has_valid_crs when an exception is thrown."""
+    crs_string = 'invalid CRS'
+    message = HarmonyMessage({'format': {'crs': crs_string}})
+    with pytest.raises(InvalidTargetCRS, match=crs_string):
+        has_valid_crs(message)
