+++ tests/unit/test_utilities.py
@@ -2,16 +2,36 @@
 
 from unittest import TestCase
 
+import numpy as np
+import pytest
 from harmony_service_lib.message import Message
+from netCDF4 import Dataset
+from numpy.testing import assert_array_equal
 
 from harmony_regridding_service.message_utilities import (
     has_valid_interpolation,
 )
-from harmony_regridding_service.utilities import get_file_mime_type
+from harmony_regridding_service.resample import (
+    _transfer_resampled_dimensions,
+)
+from harmony_regridding_service.utilities import (
+    _clone_variables,
+    _copy_var_with_attrs,
+    _copy_var_without_metadata,
+    _get_bounds_var,
+    _get_variable,
+    _integer_like,
+    _transfer_metadata,
+    _walk_groups,
+    get_file_mime_type,
+)
 
 
 class TestUtilities(TestCase):
-    """A class testing the harmony_regridding_service.utilities module."""
+    """A class testing the harmony_regridding_service.utilities module.
+
+    TODO: Update this to pytest.
+    """
 
     def test_get_file_mime_type(self):
         """Ensure a MIME type can be retrieved from an input file path."""
@@ -53,3 +73,171 @@ class TestUtilities(TestCase):
         with self.subTest('Unexpected interpolation returns False'):
             test_message = Message({'format': {'interpolation': 'Bilinear'}})
             self.assertFalse(has_valid_interpolation(test_message))
+
+
+def test__walk_groups(test_file):
+    """Demonstrate traversing all groups."""
+    target_path = test_file
+    groups = ['/a/nested/group', '/b/another/deeper/group2']
+    expected_visited = {'a', 'nested', 'group', 'b', 'another', 'deeper', 'group2'}
+
+    with Dataset(target_path, mode='w') as target_ds:
+        for group in groups:
+            target_ds.createGroup(group)
+
+    actual_visited = set()
+    with Dataset(target_path, mode='r') as validate:
+        for groups in _walk_groups(validate):
+            for group in groups:
+                actual_visited.update([group.name])
+
+    assert expected_visited == actual_visited
+
+
+def test__transfer_metadata(test_file, test_1D_dimensions_ncfile):
+    """Tests to ensure root and group level metadata is transfered to target."""
+    _generate_test_file = test_file
+
+    # metadata Set in the test 1D file
+    expected_root_metadata = {
+        'root-attribute1': 'value1',
+        'root-attribute2': 'value2',
+    }
+    expected_root_groups = {'level1-nested1', 'level1-nested2'}
+    expected_nested_metadata = {'level2-nested1': 'level2-nested1-value1'}
+
+    with (
+        Dataset(test_1D_dimensions_ncfile, mode='r') as source_ds,
+        Dataset(_generate_test_file, mode='w') as target_ds,
+    ):
+        _transfer_metadata(source_ds, target_ds)
+
+    with Dataset(_generate_test_file, mode='r') as validate:
+        root_metadata = {attr: validate.getncattr(attr) for attr in validate.ncattrs()}
+        root_groups = set(validate.groups.keys())
+        nested_group = validate['/level1-nested1/level2-nested1']
+        nested_metadata = {
+            attr: nested_group.getncattr(attr) for attr in nested_group.ncattrs()
+        }
+
+        assert expected_root_groups == root_groups
+        assert expected_root_metadata == root_metadata
+        assert expected_nested_metadata == nested_metadata
+
+
+@pytest.mark.parametrize(
+    'int_type',
+    [
+        np.byte,
+        np.ubyte,
+        np.short,
+        np.ushort,
+        np.intc,
+        np.uintc,
+        np.int_,
+        np.uint,
+        np.longlong,
+        np.ulonglong,
+    ],
+)
+def test__integer_like(int_type):
+    assert _integer_like(int_type) is True
+
+
+@pytest.mark.parametrize('float_type', [np.float16, np.float32, np.float64])
+def test__integer_like_false(float_type):
+    assert _integer_like(float_type) is False
+
+
+def test__integer_like_string():
+    assert _integer_like(str) is False
+
+
+def test__copy_var_with_attrs(
+    test_file, test_area_fxn, test_1D_dimensions_ncfile, var_info_fxn
+):
+    target_file = test_file
+    target_area = test_area_fxn()
+    var_info = var_info_fxn(test_1D_dimensions_ncfile)
+    expected_metadata = {'units': 'widgets per month'}
+    with (
+        Dataset(test_1D_dimensions_ncfile, mode='r') as source_ds,
+        Dataset(target_file, mode='w') as target_ds,
+    ):
+        _transfer_resampled_dimensions(source_ds, target_ds, target_area, var_info)
+        _copy_var_with_attrs(source_ds, target_ds, '/data')
+
+    with Dataset(target_file, mode='r') as validate:
+        actual_metadata = {
+            attr: validate['/data'].getncattr(attr)
+            for attr in validate['/data'].ncattrs()
+        }
+        assert actual_metadata == expected_metadata
+
+
+def test__copy_vars_without_metadata(
+    test_file, test_area_fxn, test_1D_dimensions_ncfile, var_info_fxn
+):
+    target_file = test_file
+    target_area = test_area_fxn()
+    var_info = var_info_fxn(test_1D_dimensions_ncfile)
+    with (
+        Dataset(test_1D_dimensions_ncfile, mode='r') as source_ds,
+        Dataset(target_file, mode='w') as target_ds,
+    ):
+        _transfer_resampled_dimensions(source_ds, target_ds, target_area, var_info)
+        _copy_var_without_metadata(source_ds, target_ds, '/data')
+
+    with Dataset(target_file, mode='r') as validate:
+        actual_metadata = {
+            attr: validate['/data'].getncattr(attr)
+            for attr in validate['/data'].ncattrs()
+        }
+        assert {} == actual_metadata
+
+
+def test__clone_variables(
+    test_file, var_info_fxn, test_area_fxn, test_1D_dimensions_ncfile
+):
+    target_file = test_file
+    var_info = var_info_fxn(test_1D_dimensions_ncfile)
+    width = 36
+    height = 18
+
+    _generate_test_area = test_area_fxn(width=width, height=height)
+
+    copy_vars = {'/time', '/time_bnds'}
+    with (
+        Dataset(test_1D_dimensions_ncfile, mode='r') as source_ds,
+        Dataset(target_file, mode='w') as target_ds,
+    ):
+        _transfer_resampled_dimensions(
+            source_ds, target_ds, _generate_test_area, var_info
+        )
+
+        copied = _clone_variables(source_ds, target_ds, copy_vars)
+
+        assert copy_vars == copied
+
+        with Dataset(target_file, mode='r') as validate:
+            assert_array_equal(validate['time_bnds'], source_ds['time_bnds'])
+            assert_array_equal(validate['time'], source_ds['time'])
+
+
+def test__get_variable(test_ATL14_ncfile):
+    with Dataset(test_ATL14_ncfile, mode='r') as source_ds:
+        var_grouped = _get_variable(source_ds, '/tile_stats/RMS_data')
+        expected_grouped = source_ds['tile_stats'].variables['RMS_data']
+        assert expected_grouped == var_grouped
+
+        var_flat = _get_variable(source_ds, '/ice_area')
+        expected_flat = source_ds.variables['ice_area']
+        assert expected_flat == var_flat
+
+
+def test__get_bounds_var(var_info_fxn, test_IMERG_ncfile):
+    var_info = var_info_fxn(test_IMERG_ncfile)
+    expected_bounds = 'lon_bnds'
+
+    actual_bounds = _get_bounds_var(var_info, '/Grid/lon')
+    assert expected_bounds == actual_bounds
